Added Exception handling

+++ Settings/BusinessLayer/Services/SettingsService.cs
@@ -1,9 +1,10 @@
 ﻿using AutoMapper;
 using BusinessLayer.Contracts;
 using BusinessLayer.Dtos;
+using BusinessLayer.Exceptions;
 using DataAccessLayer.Entities;
 using DataAccessLayer.Repositories;
-
+using static BusinessLayer.Exceptions.ApiException;
 
 namespace BusinessLayer.Services
 {
@@ -20,6 +21,10 @@ namespace BusinessLayer.Services
         }
         public async Task AddSettingAsync(SettingDto settingDto,int userId, CancellationToken token)
         {
+            if (await _settingsRepository.IsEntityExistsAsync(userId.ToString(), token))
+            {
+                throw new ApiException("Invalid operation", ExceptionStatus.BadRequest);
+            }
             var setting = _mapper.Map<SettingsInfo>(settingDto);
             setting.Id = userId.ToString();
             await _settingsRepository.CreateAsync(setting,token);
@@ -27,18 +32,30 @@ namespace BusinessLayer.Services
 
         public async Task DeleteSettingAsync(int id, CancellationToken token)
         {
+            if (!(await _settingsRepository.IsEntityExistsAsync(id.ToString(),token)))
+            {
+                throw new ApiException("Invalid operation", ExceptionStatus.BadRequest);
+            }
             await _settingsRepository.DeleteAsync(id.ToString(), token);
         }
 
         public async Task<SettingDto> GetSettingAsync(int id, CancellationToken token)
         {
             var setting = await _settingsRepository.GetByIdAsync(id.ToString(),token);
+            if (setting is null)
+            {
+                throw new ApiException("Invalid operation", ExceptionStatus.BadRequest);
+            }
 
             return _mapper.Map<SettingDto>(setting);
         }
 
         public async Task UpdateSettingAsync(SettingDto settingDto, int userId, CancellationToken token)
         {
+            if (!(await _settingsRepository.IsEntityExistsAsync(userId.ToString(), token)))
+            {
+                throw new ApiException("Invalid operation", ExceptionStatus.BadRequest);
+            }
             var setting = _mapper.Map<SettingsInfo>(settingDto);
             setting.Id = userId.ToString();
             await _settingsRepository.UpdateAsync(setting,token);
